feat: further release automation
